+++ lldb/source/Plugins/Language/ObjC/Cocoa.cpp
@@ -1040,7 +1040,7 @@ public:
 
   uint64_t CalculateNumChildren() override { return 0; }
 
-  lldb::ValueObjectSP GetChildAtIndex(size_t idx) override {
+  lldb::ValueObjectSP GetChildAtIndex(uint64_t idx) override {
     return lldb::ValueObjectSP();
   }
